requirements.txt
google-generativeai==0.3.1
Flask==2.2.2